small refactoring and added some fixed normalization values
