+++ pcse_gym/utils/normalization.py
@@ -149,7 +149,8 @@ class NormalizeMeasureObservations:
                  /,
                  has_random=False,
                  no_weather=False,
-                 loc=(52, 5.5),
+                 loc='NL',
+                 start_type='sowing',
                  mask_binary=False,
                  placeholder=-1.11,
                  reward_div=600,
@@ -159,6 +160,8 @@ class NormalizeMeasureObservations:
         self.crop_features = crop_features
         self.is_clipped = is_clipped
         self.has_random = has_random
+        self.loc = loc
+        self.start_type = start_type
 
         if not self.has_random:
             assert self.crop_features == ["DVS", "TAGP", "LAI", "NuptakeTotal", "NAVAIL", "SM"]
@@ -176,8 +179,8 @@ class NormalizeMeasureObservations:
         # elif (55.0, 23.5) in loc or (55.0, 24.0) in loc or (55.5, 23.5) in loc:
         #     loc_str = 'LT'
 
-        self.fixed_means = np.array(self.get_fixed_vals(loc, 'means'))
-        self.fixed_std = np.array(self.get_fixed_vals(loc, 'std'))
+        self.fixed_means = np.array(self.get_fixed_vals('means'))
+        self.fixed_std = np.array(self.get_fixed_vals('std'))
         self.reward_div = reward_div
 
     def normalize_measure_obs(self, obs, measure):
@@ -223,11 +226,11 @@ class NormalizeMeasureObservations:
     def unnormalize_rew(self, reward):
         return reward*self.reward_div
 
-    def get_fixed_vals(self, loc, which):
+    def get_fixed_vals(self, which):
         if which == 'means':
-            vals = self.get_means_vector(loc)
+            vals = self.get_means_vector()
         elif which == 'std':
-            vals = self.get_std_vector(loc)
+            vals = self.get_std_vector()
         else:
             raise Exception(f'Error! {which} not a valid choice')
 
@@ -245,76 +248,110 @@ class NormalizeMeasureObservations:
             case _:
                 raise Exception(f'Location error! Normalization not implemented for {loc}')
 
-    @staticmethod
-    def get_means_vector(loc):
-        if loc == 'NL':
-            return   [0.6246465476467645,
-                      4865.136421503844,
-                      1.234561809311616,
-                      150.7396483914316,
-                      161.29929450042695,
-                      0.27243507521622345,
-                      0.0,
-                      1.21883597e+07, 4.95082017e+00, 1.90003949e-01,
-                      1.22270535e+07, 4.90367558e+00, 1.95953827e-01,
-                      1.23226094e+07, 4.91493013e+00, 2.00468408e-01,
-                      1.21937849e+07, 5.04890948e+00, 2.12458688e-01,
-                      1.24103797e+07, 5.08396719e+00, 2.08391555e-01,
-                      1.25560207e+07, 5.19116343e+00, 2.00330194e-01,
-                      1.26053463e+07, 5.34218712e+00, 2.03171324e-01]
-        elif loc == 'LT':
-            return [0.33344140142967654,
-                      2675.772872654855,
-                      1.092753787112243,
-                      57.55634087558257,
-                      387.7470016444115,
-                      0.30791477941701645,
-                      0.0,
-                      9.75607378e+06, 4.36845453e-01, 2.12145007e-01,
-                      9.83234895e+06, 5.56896332e-01, 1.99204579e-01,
-                      9.95673097e+06, 5.62800509e-01, 1.91617766e-01,
-                      1.00325419e+07, 5.50421878e-01, 2.10531906e-01,
-                      1.00911787e+07, 5.02440110e-01, 1.96787789e-01,
-                      9.92729065e+06, 5.40256519e-01, 1.95853933e-01,
-                      9.93554166e+06, 5.95882976e-01, 2.01239347e-01]
-        else:
-            raise Exception(f'Error! "{loc}" not a valid choice!')
+    def get_means_vector(self):
+        match self.loc, self.start_type:
+            case 'NL', 'emergence':
+                return [0.6246465476467645,
+                        4865.136421503844,
+                        1.234561809311616,
+                        150.7396483914316,
+                        161.29929450042695,
+                        0.27243507521622345,
+                        0.0,
+                        1.21883597e+07, 4.95082017e+00, 1.90003949e-01,
+                        1.22270535e+07, 4.90367558e+00, 1.95953827e-01,
+                        1.23226094e+07, 4.91493013e+00, 2.00468408e-01,
+                        1.21937849e+07, 5.04890948e+00, 2.12458688e-01,
+                        1.24103797e+07, 5.08396719e+00, 2.08391555e-01,
+                        1.25560207e+07, 5.19116343e+00, 2.00330194e-01,
+                        1.26053463e+07, 5.34218712e+00, 2.03171324e-01]
+            case 'NL', 'sowing':
+                return [3.88126166e-01,
+                        3.91054054e+03,
+                        2.06089509e+00,
+                        1.09885505e+02,
+                        3.16524752e+02,
+                        2.83770363e-01,
+                        0.0,
+                        9.20855654e+06, 4.28197561e+00, 2.09696231e-01,
+                        9.29468958e+06, 4.28358537e+00, 1.95185809e-01,
+                        9.37642572e+06, 4.20525277e+00, 2.02279157e-01,
+                        9.33309534e+06, 4.15982483e+00, 2.09109978e-01,
+                        9.23491353e+06, 4.22654767e+00, 2.14588692e-01,
+                        9.34160976e+06, 4.18853659e+00, 2.22172062e-01,
+                        9.44925055e+06, 4.27064745e+00, 2.16439246e-01]
+            case 'LT', 'sowing':
+                return [0.33344140142967654,
+                        2675.772872654855,
+                        1.092753787112243,
+                        57.55634087558257,
+                        387.7470016444115,
+                        0.30791477941701645,
+                        0.0,
+                        9.75607378e+06, 4.36845453e-01, 2.12145007e-01,
+                        9.83234895e+06, 5.56896332e-01, 1.99204579e-01,
+                        9.95673097e+06, 5.62800509e-01, 1.91617766e-01,
+                        1.00325419e+07, 5.50421878e-01, 2.10531906e-01,
+                        1.00911787e+07, 5.02440110e-01, 1.96787789e-01,
+                        9.92729065e+06, 5.40256519e-01, 1.95853933e-01,
+                        9.93554166e+06, 5.95882976e-01, 2.01239347e-01]
+            case 'LT', 'emergence':
+                raise NotImplementedError
+            case _:
+                raise Exception(f'Error! "{self.loc} and {self.start_type}" not a valid choice!')
 
-    @staticmethod
-    def get_std_vector(loc):
-        if loc == 'NL':
-            return  [0.5728143239654657,
-              6177.7736977069735,
-              1.5758559066832276,
-              177.0009703071452,
-              91.93445733139745,
-              0.0341784277318637,
-              10.0,
-              1.21883597e+07, 4.95082017e+00, 1.90003949e-01,
-              1.22270535e+07, 4.90367558e+00, 1.95953827e-01,
-              1.23226094e+07, 4.91493013e+00, 2.00468408e-01,
-              1.21937849e+07, 5.04890948e+00, 2.12458688e-01,
-              1.24103797e+07, 5.08396719e+00, 2.08391555e-01,
-              1.25560207e+07, 5.19116343e+00, 2.00330194e-01,
-              1.26053463e+07, 5.34218712e+00, 2.03171324e-01]
-        elif loc == 'LT':
-            return [0.5577426661192912,
-              5390.806038545118,
-              1.8315136788380577,
-              109.21366379879503,
-              204.46094641439367,
-              0.021277502711722556,
-              10.0,
-              8.33064223e+06, 8.43375544e+00, 3.59799387e-01,
-              8.35175786e+06, 8.38202910e+00, 3.10559526e-01,
-              8.55516035e+06, 8.23049704e+00, 3.17290068e-01,
-              8.52174349e+06, 8.30084969e+00, 3.56926711e-01,
-              8.51462773e+06, 8.36266632e+00, 3.22802387e-01,
-              8.31931600e+06, 8.54765225e+00, 3.30367139e-01,
-              8.36117416e+06, 8.61383742e+00, 3.40634577e-01]
-        else:
-            raise Exception(f'Error! "{loc}" not a valid choice!')
 
+    def get_std_vector(self):
+        match self.loc, self.start_type:
+            case 'NL', 'emergence':
+                return [0.5728143239654657,
+                        6177.7736977069735,
+                        1.5758559066832276,
+                        177.0009703071452,
+                        91.93445733139745,
+                        0.0341784277318637,
+                        10.0,
+                        1.21883597e+07, 4.95082017e+00, 1.90003949e-01,
+                        1.22270535e+07, 4.90367558e+00, 1.95953827e-01,
+                        1.23226094e+07, 4.91493013e+00, 2.00468408e-01,
+                        1.21937849e+07, 5.04890948e+00, 2.12458688e-01,
+                        1.24103797e+07, 5.08396719e+00, 2.08391555e-01,
+                        1.25560207e+07, 5.19116343e+00, 2.00330194e-01,
+                        1.26053463e+07, 5.34218712e+00, 2.03171324e-01]
+            case 'NL', 'sowing':
+                return [5.59906750e-01,
+                        6.54310379e+03,
+                        2.60923433e+00,
+                        1.54730097e+02,
+                        1.61885806e+02,
+                        3.26634375e-02,
+                        10.0,
+                        7.45039155e+06, 5.29834422e+00, 3.19931624e-01,
+                        7.43363615e+06, 5.31585818e+00, 3.07920195e-01,
+                        7.50510408e+06, 5.34892744e+00, 3.40834363e-01,
+                        7.57753950e+06, 5.44657448e+00, 3.34261841e-01,
+                        7.49673622e+06, 5.45726454e+00, 3.29862859e-01,
+                        7.49611165e+06, 5.42787962e+00, 3.47691447e-01,
+                        7.60912059e+06, 5.40121932e+00, 3.49741439e-01]
+            case 'LT', 'sowing':
+                return [0.5577426661192912,
+                        5390.806038545118,
+                        1.8315136788380577,
+                        109.21366379879503,
+                        204.46094641439367,
+                        0.021277502711722556,
+                        10.0,
+                        8.33064223e+06, 8.43375544e+00, 3.59799387e-01,
+                        8.35175786e+06, 8.38202910e+00, 3.10559526e-01,
+                        8.55516035e+06, 8.23049704e+00, 3.17290068e-01,
+                        8.52174349e+06, 8.30084969e+00, 3.56926711e-01,
+                        8.51462773e+06, 8.36266632e+00, 3.22802387e-01,
+                        8.31931600e+06, 8.54765225e+00, 3.30367139e-01,
+                        8.36117416e+06, 8.61383742e+00, 3.40634577e-01]
+            case 'LT', 'emergence':
+                raise NotImplementedError
+            case _:
+                raise Exception(f'Error! "{self.loc} and {self.start_type}" not a valid choice!')
 
 class RunningReward:
     '''
